backend/requirements.txt
pandas==1.5.3
requests==2.28.2
requests-cache==1.1.0 
sanic[ext]
SQLAlchemy==1.4.44
PyMySQL[rsa]
PyMySQL==1.0.2
python-dotenv==0.21.0
bcrypt==4.0.1
pyjwt==2.6.0
pytz==2021.1
apscheduler==3.9.1
yagmail==0.15.285
google-cloud-secret-manager
ldap3==2.9.1